Refactor package installation and service management
This commit refactors the package installation and service management in the `converge.yml` and `inet-http.yml` files. It updates the syntax of the `ansible.builtin.apt` and `ansible.builtin.package` tasks to use YAML format and adds explicit state and name parameters. Additionally, it removes the `curl-minimal` package before installing `curl` and ensures that the Apache service is stopped and disabled after installation. These changes improve the readability and maintainability of the code.